+++ src/vsg/vk/PushConstants.cpp
@@ -42,6 +42,6 @@ void PushConstants::dispatch(CommandBuffer& commandBuffer) const
     if (pipelineLayout)
     {
 
-        vkCmdPushConstants(commandBuffer, *pipelineLayout, _stageFlags, _offset, _data->dataSize(), _data->dataPointer());
+        vkCmdPushConstants(commandBuffer, *pipelineLayout, _stageFlags, _offset, static_cast<uint32_t>(_data->dataSize()), _data->dataPointer());
     }
 }
